Rename based on `flutter_lint`
