+++ lib/easy_sidemenu.dart
@@ -1,6 +1,6 @@
 library easy_sidemenu;
 
-export 'package:easy_sidemenu/src/SideMenu.dart';
-export 'package:easy_sidemenu/src/SideMenuItem.dart';
-export 'package:easy_sidemenu/src/SideMenuDisplayMode.dart';
-export 'package:easy_sidemenu/src/SideMenuStyle.dart';
+export 'package:easy_sidemenu/src/side_menu.dart';
+export 'package:easy_sidemenu/src/side_menu_item.dart';
+export 'package:easy_sidemenu/src/side_menu_display_mode.dart';
+export 'package:easy_sidemenu/src/side_menu_style.dart';
