+++ lib/src/Global/Global.dart
@@ -1,5 +1,5 @@
-import '../SideMenuDisplayMode.dart';
-import '../SideMenuStyle.dart';
+import 'package:easy_sidemenu/src/side_menu_display_mode.dart';
+import 'package:easy_sidemenu/src/side_menu_style.dart';
 import 'package:flutter/widgets.dart';
 
 class Global {
