+++ lib/src/side_menu.dart
@@ -1,9 +1,8 @@
+import 'package:easy_sidemenu/src/global/global.dart';
+import 'package:easy_sidemenu/src/side_menu_display_mode.dart';
+import 'package:easy_sidemenu/src/side_menu_item.dart';
+import 'package:easy_sidemenu/src/side_menu_style.dart';
 import 'package:flutter/material.dart';
-
-import 'SideMenuStyle.dart';
-import 'Global/Global.dart';
-import 'SideMenuItem.dart';
-import 'SideMenuDisplayMode.dart';
 
 class SideMenu extends StatelessWidget {
   /// Page controller to control [PageView] widget
